# optimize the target list to be scrapped