--- a/scrape/scrape_test.go
+++ b/scrape/scrape_test.go
@@ -100,8 +100,8 @@ func TestDroppedTargetsList(t *testing.T) {
 	if len(sp.droppedTargets) != expectedLength {
 		t.Fatalf("Length of dropped targets exceeded expected length, expected %v, got %v", expectedLength, len(sp.droppedTargets))
 	}
-	if sp.droppedTargets[0].DiscoveredLabels().String() != expectedLabelSetString {
-		t.Fatalf("Got %v, expected %v", sp.droppedTargets[0].DiscoveredLabels().String(), expectedLabelSetString)
+	if sp.DroppedTargets()[0].DiscoveredLabels().String() != expectedLabelSetString {
+		t.Fatalf("Got %v, expected %v", sp.DroppedTargets()[0].DiscoveredLabels().String(), expectedLabelSetString)
 	}
 }