@Nexucis 2022-05-03
>  Error: File is not `gofumpt`-ed with `-extra` (gofumpt)

I don't get why it is mad at me. 
`make common-format` doesn't change anything on my side. 
Same when I'm running `gofumpt -w scrape/target.go` like it is suggests when you are searching on google how to fix this issue.