@roidelapluie 2022-05-03
gofumpt -w scrape/target.go -extra