@Nexucis 2022-05-03
> --- FAIL: TestQueryLog/rule_queries,_[::1]:50285,_enabled_at_start (6.58s)

I don't see why this test is failing. Is it related to the changes ?