@Nexucis 2022-05-10
Soooo,

Dropping the `LabelsBuilder` in favor of a map usage seems to work a bit. We gain 1 ~2 VCPUs on 8.
On our side, having a logic to drop the target and so avoiding the relabelling doesn't reduce the CPU used. But it's probably because in Amadeus, our clusters are too big :(.

I will try to apply the same optimization on the method `applyRelabelling`. Perhaps we can still gain some CPU usage there.

![image](https://user-images.githubusercontent.com/4548045/167661262-e7d4552b-429b-48af-85e9-28a7654e83ff.png)

But even, if the optimization is not sufficient, I think it will be ok to go ahead with this PR and to merge it.

I fear that to be finally be able to solve this issue, we will need to allow the possibility to receive the k8s event and to react on it. That doesn't mean it should replace the current logic, I'm just saying we could have 2 enters to refresh the targets:
- the current one that reload everything from scratch
- one entry that gives just the update of the target that changed