@Nexucis 2022-05-12
this PR is ready to be reviewed. At this point I don't know what we could improve more than what I found and fixed. Of course, someone with a better understanding about how Go is working and the SD is working can have better ideas.

I still think this PR can be merged as it shown in a real production environment a gain about CPUs usage. But not enough to fix the issue.

Probably a combination with this PR and https://github.com/prometheus/prometheus/pull/10634 can help to mitigate the issue.

I'm afraid the only real fix is to have another channel in the SD that will be used to receive a unique update of a target. In the particular case of the k8s SD it will be used to FW the update of the PODs (for example). 

@roidelapluie or @brancz can you review this PR please ? I think it can be cool to be able to merge it before the release.