@Nexucis 2022-06-03
I run a simple benchmark on the `Sync` method and here the result between the main branch (old time) and this PR (new time)

```txt
go test -bench=Sync -count 5 -run=^# -benchmem | tee old.txt
goos: darwin
goarch: amd64
pkg: github.com/prometheus/prometheus/scrape
cpu: Intel(R) Core(TM) i5-1038NG7 CPU @ 2.00GHz
BenchmarkScrapePool_Sync-8   	     192	   6143772 ns/op	 2258378 B/op	   33804 allocs/op
BenchmarkScrapePool_Sync-8   	     194	   6325346 ns/op	 2258979 B/op	   33804 allocs/op
BenchmarkScrapePool_Sync-8   	     190	   6159117 ns/op	 2258325 B/op	   33804 allocs/op
BenchmarkScrapePool_Sync-8   	     194	   6134170 ns/op	 2258905 B/op	   33805 allocs/op
BenchmarkScrapePool_Sync-8   	     194	   6104749 ns/op	 2258227 B/op	   33803 allocs/op
PASS
ok  	github.com/prometheus/prometheus/scrape	10.031s


go test -bench=Sync -count 5 -run=^# -benchmem | tee new.txt
goos: darwin
goarch: amd64
pkg: github.com/prometheus/prometheus/scrape
cpu: Intel(R) Core(TM) i5-1038NG7 CPU @ 2.00GHz
BenchmarkScrapePool_Sync-8   	     258	   4665387 ns/op	 1571625 B/op	   24035 allocs/op
BenchmarkScrapePool_Sync-8   	     247	   4501856 ns/op	 1572085 B/op   24036 allocs/op
BenchmarkScrapePool_Sync-8   	     274	   4379673 ns/op	 1571774 B/op	   24033 allocs/op
BenchmarkScrapePool_Sync-8   	     273	   4382959 ns/op	 1571421 B/op	   24034 allocs/op
BenchmarkScrapePool_Sync-8   	     271	   4380908 ns/op	 1571462 B/op	   24034 allocs/op
PASS
ok  	github.com/prometheus/prometheus/scrape	8.670s


benchstat old.txt new.txt
name               old time/op    new time/op    delta
ScrapePool_Sync-8    6.17ms ± 2%    4.46ms ± 5%  -27.72%  (p=0.008 n=5+5)

name               old alloc/op   new alloc/op   delta
ScrapePool_Sync-8    2.26MB ± 0%    1.57MB ± 0%  -30.41%  (p=0.008 n=5+5)

name               old allocs/op  new allocs/op  delta
ScrapePool_Sync-8     33.8k ± 0%     24.0k ± 0%  -28.90%  (p=0.008 n=5+5)
```