@Nexucis 2022-06-03
and when I'm running the same benchmark with the usage of the labelsBuilder in the method `buildTargetLabels`, here the result:

```txt
go test -bench=Sync -count 5 -run=^# -benchmem | tee new_with_labels_builder.txt
goos: darwin
goarch: amd64
pkg: github.com/prometheus/prometheus/scrape
cpu: Intel(R) Core(TM) i5-1038NG7 CPU @ 2.00GHz
BenchmarkScrapePool_Sync-8   	     253	   4825677 ns/op	 2132310 B/op	   28038 allocs/op
BenchmarkScrapePool_Sync-8   	     255	   4687988 ns/op	 2132786 B/op	   28039 allocs/op
BenchmarkScrapePool_Sync-8   	     255	   4641599 ns/op	 2132695 B/op	   28039 allocs/op
BenchmarkScrapePool_Sync-8   	     258	   4635619 ns/op	 2132150 B/op	   28037 allocs/op
BenchmarkScrapePool_Sync-8   	     256	   4651736 ns/op	 2132173 B/op	   28037 allocs/op
PASS
ok  	github.com/prometheus/prometheus/scrape	9.468s

benchstat  new.txt new_with_labels_builder.txt
name               old time/op    new time/op    delta
ScrapePool_Sync-8    4.46ms ± 5%    4.69ms ± 3%     ~     (p=0.056 n=5+5)

name               old alloc/op   new alloc/op   delta
ScrapePool_Sync-8    1.57MB ± 0%    2.13MB ± 0%  +35.68%  (p=0.008 n=5+5)

name               old allocs/op  new allocs/op  delta
ScrapePool_Sync-8     24.0k ± 0%     28.0k ± 0%  +16.66%  (p=0.008 n=5+5)
```

Old is the current implementation of the PR. New is the implementation of the method `buildTargetLabels` from the commit [74c8c155bdba](https://github.com/prometheus/prometheus/pull/10662/commits/74c8c155bdba101cc75882f518243d5e2b8b1ff2#diff-f9baaa77ffadc6027a5881865a568d273ceac36a04528b6e6cb7a745bf95eac3R507-R537)