@Nexucis 2022-06-03
maybe the benchmark is too simple, I was more or less expected to see a difference in terme of CPU usage.