@Nexucis 2022-06-07
I added a suite of benchmark to see if the performances are stable when the number of targets are increasing. And it appears it is the case. 

`old.txt` is the bench running on the main branch
`new.txt` is the bench running with the last commit on this PR

```txt
go test -bench=Sync -count 10 -run=^# -benchmem | tee old.txt
goos: darwin
goarch: amd64
pkg: github.com/prometheus/prometheus/scrape
cpu: Intel(R) Core(TM) i5-1038NG7 CPU @ 2.00GHz
BenchmarkScrapePool_Sync/1k_targets-8         	     196	   6042453 ns/op	 2258344 B/op	   33803 allocs/op
BenchmarkScrapePool_Sync/1k_targets-8         	     200	   6071995 ns/op	 2258862 B/op	   33803 allocs/op
BenchmarkScrapePool_Sync/1k_targets-8         	     199	   5978866 ns/op	 2258249 B/op	   33803 allocs/op
BenchmarkScrapePool_Sync/1k_targets-8         	     199	   5937812 ns/op	 2258693 B/op	   33802 allocs/op
BenchmarkScrapePool_Sync/1k_targets-8         	     200	   6085858 ns/op	 2258162 B/op	   33802 allocs/op
BenchmarkScrapePool_Sync/1k_targets-8         	     199	   5943592 ns/op	 2258170 B/op	   33802 allocs/op
BenchmarkScrapePool_Sync/1k_targets-8         	     200	   5939149 ns/op	 2258187 B/op	   33802 allocs/op
BenchmarkScrapePool_Sync/1k_targets-8         	     201	   5935889 ns/op	 2258187 B/op	   33802 allocs/op
BenchmarkScrapePool_Sync/1k_targets-8         	     201	   5922343 ns/op	 2258183 B/op	   33802 allocs/op
BenchmarkScrapePool_Sync/1k_targets-8         	     195	   5914851 ns/op	 2258215 B/op	   33803 allocs/op
BenchmarkScrapePool_Sync/10k-8                	      19	  57927421 ns/op	22990018 B/op	  342935 allocs/op
BenchmarkScrapePool_Sync/10k-8                	      18	  56946789 ns/op	23002349 B/op	  343110 allocs/op
BenchmarkScrapePool_Sync/10k-8                	      19	  57331993 ns/op	22989989 B/op	  342936 allocs/op
BenchmarkScrapePool_Sync/10k-8                	      19	  57179699 ns/op	22989972 B/op	  342935 allocs/op
BenchmarkScrapePool_Sync/10k-8                	      19	  57661313 ns/op	22990000 B/op	  342936 allocs/op
BenchmarkScrapePool_Sync/10k-8                	      18	  57289474 ns/op	23002603 B/op	  343113 allocs/op
BenchmarkScrapePool_Sync/10k-8                	      18	  57331912 ns/op	23002320 B/op	  343111 allocs/op
BenchmarkScrapePool_Sync/10k-8                	      18	  57239766 ns/op	23002792 B/op	  343112 allocs/op
BenchmarkScrapePool_Sync/10k-8                	      19	  58867540 ns/op	22989584 B/op	  342932 allocs/op
BenchmarkScrapePool_Sync/10k-8                	      19	  57716155 ns/op	22990525 B/op	  342939 allocs/op
BenchmarkScrapePool_Sync/100k-8               	       2	 561901938 ns/op	251003236 B/op	 3699613 allocs/op
BenchmarkScrapePool_Sync/100k-8               	       2	 562174696 ns/op	251003172 B/op	 3699612 allocs/op
BenchmarkScrapePool_Sync/100k-8               	       2	 561638742 ns/op	251005832 B/op	 3699633 allocs/op
BenchmarkScrapePool_Sync/100k-8               	       2	 557759962 ns/op	251004900 B/op	 3699624 allocs/op
BenchmarkScrapePool_Sync/100k-8               	       2	 582328120 ns/op	251005844 B/op	 3699637 allocs/op
BenchmarkScrapePool_Sync/100k-8               	       2	 564987792 ns/op	251004476 B/op	 3699617 allocs/op
BenchmarkScrapePool_Sync/100k-8               	       2	 555913553 ns/op	251002396 B/op	 3699601 allocs/op
BenchmarkScrapePool_Sync/100k-8               	       2	 556632320 ns/op	251005300 B/op	 3699630 allocs/op
BenchmarkScrapePool_Sync/100k-8               	       2	 556756732 ns/op	251004508 B/op	 3699618 allocs/op
BenchmarkScrapePool_Sync/100k-8               	       2	 553843642 ns/op	251002536 B/op	 3699604 allocs/op

go test -bench=Sync -count 10 -run=^# -benchmem | tee new.txt
goos: darwin
goarch: amd64
pkg: github.com/prometheus/prometheus/scrape
cpu: Intel(R) Core(TM) i5-1038NG7 CPU @ 2.00GHz
BenchmarkScrapePool_Sync/1k_targets-8         	     283	   4177944 ns/op	 1507338 B/op	   24032 allocs/op
BenchmarkScrapePool_Sync/1k_targets-8         	     284	   4191782 ns/op	 1507739 B/op	   24033 allocs/op
BenchmarkScrapePool_Sync/1k_targets-8         	     286	   4175675 ns/op	 1507686 B/op	   24033 allocs/op
BenchmarkScrapePool_Sync/1k_targets-8         	     288	   4128414 ns/op	 1507224 B/op	   24032 allocs/op
BenchmarkScrapePool_Sync/1k_targets-8         	     286	   4159042 ns/op	 1507263 B/op	   24032 allocs/op
BenchmarkScrapePool_Sync/1k_targets-8         	     277	   4173712 ns/op	 1507317 B/op	   24033 allocs/op
BenchmarkScrapePool_Sync/1k_targets-8         	     286	   4153974 ns/op	 1507241 B/op	   24032 allocs/op
BenchmarkScrapePool_Sync/1k_targets-8         	     286	   4146969 ns/op	 1507251 B/op	   24032 allocs/op
BenchmarkScrapePool_Sync/1k_targets-8         	     289	   4141376 ns/op	 1507312 B/op	   24033 allocs/op
BenchmarkScrapePool_Sync/1k_targets-8         	     289	   4132895 ns/op	 1507221 B/op	   24032 allocs/op
BenchmarkScrapePool_Sync/10k-8                	      30	  40541423 ns/op	15187485 B/op	  242023 allocs/op
BenchmarkScrapePool_Sync/10k-8                	      30	  39555969 ns/op	15187625 B/op	  242025 allocs/op
BenchmarkScrapePool_Sync/10k-8                	      28	  40276782 ns/op	15197361 B/op	  242163 allocs/op
BenchmarkScrapePool_Sync/10k-8                	      25	  40112002 ns/op	15215336 B/op	  242418 allocs/op
BenchmarkScrapePool_Sync/10k-8                	      30	  39538454 ns/op	15187697 B/op	  242025 allocs/op
BenchmarkScrapePool_Sync/10k-8                	      30	  39757235 ns/op	15187142 B/op	  242020 allocs/op
BenchmarkScrapePool_Sync/10k-8                	      28	  39739595 ns/op	15197367 B/op	  242163 allocs/op
BenchmarkScrapePool_Sync/10k-8                	      30	  39558885 ns/op	15187311 B/op	  242021 allocs/op
BenchmarkScrapePool_Sync/10k-8                	      30	  39622751 ns/op	15187464 B/op	  242022 allocs/op
BenchmarkScrapePool_Sync/10k-8                	      30	  39728030 ns/op	15187539 B/op	  242023 allocs/op
BenchmarkScrapePool_Sync/100k-8               	       3	 404789837 ns/op	164580250 B/op	 2599909 allocs/op
BenchmarkScrapePool_Sync/100k-8               	       3	 390574865 ns/op	164580885 B/op	 2599917 allocs/op
BenchmarkScrapePool_Sync/100k-8               	       3	 390941605 ns/op	164581264 B/op	 2599922 allocs/op
BenchmarkScrapePool_Sync/100k-8               	       3	 391106574 ns/op	164580170 B/op	 2599909 allocs/op
BenchmarkScrapePool_Sync/100k-8               	       3	 404623794 ns/op	164579930 B/op	 2599908 allocs/op
BenchmarkScrapePool_Sync/100k-8               	       3	 393653526 ns/op	164580853 B/op	 2599916 allocs/op
BenchmarkScrapePool_Sync/100k-8               	       3	 388781652 ns/op	164580485 B/op	 2599913 allocs/op
BenchmarkScrapePool_Sync/100k-8               	       3	 388447932 ns/op	164580538 B/op	 2599913 allocs/op
BenchmarkScrapePool_Sync/100k-8               	       3	 391567633 ns/op	164580656 B/op	 2599915 allocs/op
BenchmarkScrapePool_Sync/100k-8               	       3	 388154674 ns/op	164580410 B/op	 2599912 allocs/op

benchstat  old.txt new.txt
name                          old time/op    new time/op    delta
ScrapePool_Sync/1k_targets-8    5.98ms ± 2%    4.16ms ± 1%  -30.43%  (p=0.000 n=10+10)
ScrapePool_Sync/10k-8           57.4ms ± 1%    39.8ms ± 2%  -30.59%  (p=0.000 n=9+10)
ScrapePool_Sync/100k-8           559ms ± 1%     390ms ± 1%  -30.17%  (p=0.000 n=9+8)

name                          old alloc/op   new alloc/op   delta
ScrapePool_Sync/1k_targets-8    2.26MB ± 0%    1.51MB ± 0%  -33.25%  (p=0.000 n=8+8)
ScrapePool_Sync/10k-8           23.0MB ± 0%    15.2MB ± 0%  -33.94%  (p=0.000 n=10+9)
ScrapePool_Sync/100k-8           251MB ± 0%     165MB ± 0%  -34.43%  (p=0.000 n=10+10)

name                          old allocs/op  new allocs/op  delta
ScrapePool_Sync/1k_targets-8     33.8k ± 0%     24.0k ± 0%  -28.90%  (p=0.000 n=10+10)
ScrapePool_Sync/10k-8             343k ± 0%      242k ± 0%  -29.43%  (p=0.000 n=10+9)
ScrapePool_Sync/100k-8           3.70M ± 0%     2.60M ± 0%  -29.72%  (p=0.000 n=10+10)
```


Do you think it's enough to go further in the review of this PR @roidelapluie ?