@bboreham 2023-03-09
Sorry I did not notice this previously.  #12048 and #12084 are changing the same code, probably for the same reasons.

I ran the benchmark added in this PR at commit 53bb046d2 and after making my changes:

```
goos: linux
goarch: amd64
pkg: github.com/prometheus/prometheus/scrape
cpu: Intel(R) Core(TM) i5-9400F CPU @ 2.90GHz
                             │    sec/op     │    sec/op     vs base               │
ScrapePool_Sync/1k_targets-4   12.359m ± 34%   8.092m ± 29%  -34.53% (p=0.008 n=5)
ScrapePool_Sync/10k-4           97.02m ± 19%   79.69m ± 12%  -17.86% (p=0.008 n=5)
ScrapePool_Sync/100k-4          962.7m ± 16%   807.6m ±  5%  -16.11% (p=0.008 n=5)

                             │     B/op     │     B/op      vs base               │
ScrapePool_Sync/1k_targets-4   2.384Mi ± 0%   1.268Mi ± 0%  -46.80% (p=0.008 n=5)
ScrapePool_Sync/10k-4          24.33Mi ± 0%   13.14Mi ± 0%  -45.98% (p=0.008 n=5)
ScrapePool_Sync/100k-4         262.3Mi ± 8%   150.9Mi ± 0%  -42.47% (p=0.008 n=5)
geomean                        24.78Mi        13.60Mi       -45.12%

                             │  allocs/op  │  allocs/op   vs base               │
ScrapePool_Sync/1k_targets-4   34.83k ± 0%   22.81k ± 0%  -34.52% (p=0.008 n=5)
ScrapePool_Sync/10k-4          354.8k ± 0%   234.0k ± 0%  -34.03% (p=0.008 n=5)
ScrapePool_Sync/100k-4         3.800M ± 8%   2.600M ± 0%  -31.58% (p=0.008 n=5)
geomean                        360.8k        240.3k       -33.39%
```

Compare to previous results [posted above](https://github.com/prometheus/prometheus/pull/10662#issuecomment-1148463560) (on a different CPU):
```
cpu: Intel(R) Core(TM) i5-1038NG7 CPU @ 2.00GHz
benchstat  old.txt new.txt
name                          old time/op    new time/op    delta
ScrapePool_Sync/1k_targets-8    5.98ms ± 2%    4.16ms ± 1%  -30.43%  (p=0.000 n=10+10)
ScrapePool_Sync/10k-8           57.4ms ± 1%    39.8ms ± 2%  -30.59%  (p=0.000 n=9+10)
ScrapePool_Sync/100k-8           559ms ± 1%     390ms ± 1%  -30.17%  (p=0.000 n=9+8)

name                          old alloc/op   new alloc/op   delta
ScrapePool_Sync/1k_targets-8    2.26MB ± 0%    1.51MB ± 0%  -33.25%  (p=0.000 n=8+8)
ScrapePool_Sync/10k-8           23.0MB ± 0%    15.2MB ± 0%  -33.94%  (p=0.000 n=10+9)
ScrapePool_Sync/100k-8           251MB ± 0%     165MB ± 0%  -34.43%  (p=0.000 n=10+10)

name                          old allocs/op  new allocs/op  delta
ScrapePool_Sync/1k_targets-8     33.8k ± 0%     24.0k ± 0%  -28.90%  (p=0.000 n=10+10)
ScrapePool_Sync/10k-8             343k ± 0%      242k ± 0%  -29.43%  (p=0.000 n=10+9)
ScrapePool_Sync/100k-8           3.70M ± 0%     2.60M ± 0%  -29.72%  (p=0.000 n=10+10)
```

So it looks like I'm doing better on memory allocation bytes but I didn't get such a good speedup on larger syncs. 
I will note that 100K targets in one group is not what most people experience - for instance Kubernetes generates one group for each pod.